+++ hdx_hapi/endpoints/get_conflict_events.py
@@ -32,22 +32,18 @@ CONFIG = get_config()
 router = APIRouter(
     tags=['Coordination & Context'],
 )
-
 SUMMARY_TEXT = 'Get the list of conflict events'
 
+ROUTER_DICT = {
+    'response_model': HapiGenericResponse[ConflictEventResponse],
+    'summary': SUMMARY_TEXT,
+    'responses': ERROR_RESPONSES,
+}
 
-@router.get(
-    '/api/coordination-context/conflict-event',
-    response_model=HapiGenericResponse[ConflictEventResponse],
-    summary=SUMMARY_TEXT,
-    include_in_schema=False,
-)
-@router.get(
-    '/api/v1/coordination-context/conflict-event',
-    response_model=HapiGenericResponse[ConflictEventResponse],
-    responses=ERROR_RESPONSES,  # type: ignore
-    summary=SUMMARY_TEXT,
-)
+
+@router.get('/api/coordination-context/conflict-events', include_in_schema=False, **ROUTER_DICT)
+@router.get('/api/v1/coordination-context/conflict-event', include_in_schema=False, **ROUTER_DICT)
+@router.get('/api/v2/coordination-context/conflict-events', **ROUTER_DICT)
 async def get_conflict_event(
     # ref_period_parameters: Annotated[ReferencePeriodParameters, Depends(reference_period_parameters)],
     common_date_range_params: Annotated[CommonDateRangeParams, Depends(common_date_range_params)],