+++ hdx_hapi/endpoints/get_food_security.py
@@ -30,22 +30,21 @@ from hdx_hapi.services.sql_alchemy_session import get_db
 
 CONFIG = get_config()
 router = APIRouter(
-    tags=['Food Security & Nutrition'],
+    tags=['Food Security, Nutrition & Poverty'],
 )
 
+SUMMARY_TEXT = 'Get food security data'
 
-@router.get(
-    '/api/food/food-security',
-    response_model=HapiGenericResponse[FoodSecurityResponse],
-    summary='Get food security data',
-    include_in_schema=False,
-)
-@router.get(
-    '/api/v1/food/food-security',
-    response_model=HapiGenericResponse[FoodSecurityResponse],
-    responses=ERROR_RESPONSES,  # type: ignore
-    summary='Get food security data',
-)
+ROUTER_DICT = {
+    'response_model': HapiGenericResponse[FoodSecurityResponse],
+    'summary': SUMMARY_TEXT,
+    'responses': ERROR_RESPONSES,
+}
+
+
+@router.get('/api/food-security-nutrition-poverty/food-security', include_in_schema=False, **ROUTER_DICT)
+@router.get('/api/v1/food/food-security', include_in_schema=False, **ROUTER_DICT)
+@router.get('/api/v2/food-security-nutrition-poverty/food-security', **ROUTER_DICT)
 async def get_food_security(
     common_date_range_params: Annotated[CommonDateRangeParams, Depends(common_date_range_params)],
     common_location_params: Annotated[CommonLocationParameters, Depends(common_location_parameters)],