+++ tests/test_endpoints/test_food_price_endpoint.py
@@ -13,7 +13,7 @@ from tests.util.util import split_items_by_admin_level
 
 log = logging.getLogger(__name__)
 
-ENDPOINT_ROUTER = '/api/v1/food/food-price'
+ENDPOINT_ROUTER = '/api/v2/food-security-nutrition-poverty/food-prices-market-monitor'
 endpoint_data = endpoint_data[ENDPOINT_ROUTER]
 query_parameters = endpoint_data['query_parameters']
 expected_fields = endpoint_data['expected_fields']