+++ docs/changelog.md
@@ -17,6 +17,22 @@ the API.
 
 ## Latest changes
 
+### Unreleased
+- Aligned category and sub-category endpoint names in the docs with
+  [HDX data grids](https://data.humdata.org/dashboards/overview-of-data-grids?)
+  - affected-people/refugees -> affected-people/refugees-persons-of-concern
+  - population-social/population -> geography-infrastructure/baseline-population
+  - coordination-context/conflict-event -> coordination-context/conflict-events
+  - food/food-security -> food-security-nutrition-poverty/food-security
+  - food/food-price -> food-security-nutrition-poverty/food-prices-market-monitor
+  - population-social/poverty-rate -> food-security-nutrition-poverty/poverty-rate
+  - population-social/population -> geography-infrastructure/baseline-population
+- Note that the new endpoints are being released under v2 of the api and 
+  previous versions are maintained under v1
+
+### 2025-01-06
+- Added functionality to filter by time period
+
 ### 2024-12-19
 
 - Increased temporal coverage for humanitarian needs